+++ src/back/dao/AccountDAO.java
@@ -1,0 +1,21 @@
+package back.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+
+public class AccountDAO {
+    Connection conn = null;
+    PreparedStatement pt = null;
+    ResultSet rs = null;
+
+//    유저 정보 수정
+    public void modifyUserInfo() {
+
+    }
+
+//    회원 탈퇴
+    public void withdrawlUserAccount() {
+
+    }
+}
